+++ src/app/compare/CompareItem.tsx
@@ -12,6 +12,7 @@ import { ColumnDefinition, Row, TableContext } from 'app/organizer/table-types';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
 import { noop } from 'app/utils/functions';
 import { useSetCSSVarToHeight, useShiftHeld } from 'app/utils/hooks';
+import { nonPullablePostmasterItem } from 'app/utils/item-utils';
 import clsx from 'clsx';
 import { memo, useCallback, useMemo, useRef } from 'react';
 import { useSelector } from 'react-redux';
@@ -23,6 +24,7 @@ import {
   faAngleLeft,
   faAngleRight,
   faArrowCircleDown,
+  faExclamationTriangle,
   shoppingCart,
 } from '../shell/icons';
 import styles from './CompareItem.m.scss';
@@ -62,6 +64,12 @@ export default memo(function CompareItem({
         <div className={styles.itemActions}>
           {item.vendor ? (
             <VendorItemWarning item={item} />
+          ) : nonPullablePostmasterItem(item) ? (
+            <PressTip elementType="span" tooltip={t('MovePopup.CantPullFromPostmaster')}>
+              <ActionButton onClick={noop} disabled>
+                <AppIcon icon={faExclamationTriangle} />
+              </ActionButton>
+            </PressTip>
           ) : (
             <ActionButton title={t('Hotkey.Pull')} onClick={pullItem}>
               <AppIcon icon={faArrowCircleDown} />